Move skill level class into the new FieldPopulators folder, make it a module and include in ApplDraft :hammer:
